testing and debugging coarse, loadbalance and communicate

+++ include/Class_Para_Tree_3D.tpp
@@ -3860,7 +3860,7 @@ private:
 			//update max_depth
 			error_flag = MPI_Allreduce(&octree.local_max_depth,&max_depth,1,MPI_UINT8_T,MPI_MAX,MPI_COMM_WORLD);
 			//update global_num_octants
-			uint64_t local_num_octants = octree.getNumOctants();
+			uint64_t local_num_octants = (uint64_t) octree.getNumOctants();
 			error_flag = MPI_Allreduce(&local_num_octants,&global_num_octants,1,MPI_UINT64_T,MPI_SUM,MPI_COMM_WORLD);
 			//update partition_range_globalidx
 			uint64_t rbuff [nproc];
